MetaManager Creation and Update
Tonight, we worked on create meta manager. We are done with meta manager loading path function that it will connect to ConfigurationManager and read the path information. It will read and store the meta information into meta_log under the drive. Also, we have two drives enables.

Besides that, we wrote function such that it will update the strikes if necessarily. Everyday the user request stoculus. It will update the strikes into the meta file.

Careful, we have "_ " protected filter in the naming file. So we are not expecting any ticker starts with "_"